+++ src/server/routes/delete.js
